Separate cache directory from data directory
Previously both the database file and the converted files were stored in
the data directory. From now on the converted files will be stored in
the cache directory.

Manual actions required:
- update your config by adding the new configuration key
- remove `thumbnails` and `tracks` from your data directory

It is better to remove this bandaid early as this is a change that is
not backwards compatibile and I belive that introducing it in a backward
compatibile way at this stage in development can be ignored as this
software has only two users as far as I know (myself included).